Use of socket ID to indentify user instead of username
As Username can potentially be changed (example: logs in as a normal users name, then sends he's a mod to socket.off and he just shutted down the server without mod on prior versions). Socket ID can also be manipulated but one's exact is unkown to the others and it is harder to manipulate it.